Remove sglang/srt/openai_api
- Replace all imports with sglang/srt/entrypoints/openai
- Remove batch and file endpoints in http_server.py
- Update openai-compatible endpoints in http_server.py
